+++ packages/expo-cli/src/commands/start/TerminalUI.ts
@@ -15,10 +15,10 @@ import {
   Webpack,
 } from 'xdl';
 
-import { loginOrRegisterIfLoggedOutAsync } from '../../accounts';
 import Log from '../../log';
 import { selectAsync } from '../../utils/prompts';
 import { learnMore } from '../utils/TerminalLink';
+import { loginOrRegisterIfLoggedOutAsync } from '../utils/accounts';
 import { openInEditorAsync } from '../utils/openInEditorAsync';
 import urlOpts from '../utils/urlOpts';
