+++ packages/expo-cli/src/exp.ts
@@ -31,10 +31,10 @@ import {
 } from 'xdl';
 
 import { AbortCommandError, SilentError } from './CommandError';
-import { loginOrRegisterAsync } from './accounts';
 import StatusEventEmitter from './analytics/StatusEventEmitter';
 import { registerCommands } from './commands';
 import { learnMore } from './commands/utils/TerminalLink';
+import { loginOrRegisterAsync } from './commands/utils/accounts';
 import { profileMethod } from './commands/utils/profileMethod';
 import urlOpts from './commands/utils/urlOpts';
 import Log from './log';
